Add: Breadcrumb & Settings, etc. icons
