+++ src/js/constants/texts.ts
@@ -12,6 +12,7 @@ export const TEXTS = {
     SUNDAR_GUTKA_BAANI: 'Baani',
     AMRIT_KEERTAN: 'Amrit Keertan',
     MARYADA: 'Rehat Maryadha - Sikh Code of Conduct',
+    SEARCH_RESULTS: 'Search Results',
   },
   GDPR_NOTICE: `
     Our service uses cookies to remember your preferences to give a great experience. If you do not agree to SikhiToTheMax's use of cookies, please use your browser's private/incognito mode. Otherwise, by continuing the use of the service, you consent to the use of cookies. Read more <a href="/help#Web-why-does-sikhitothemax-website-use-cookies-and-loc">here</a>.
