+++ src/scss/_results.scss
@@ -1,3 +1,12 @@
+.search-results-display {
+  margin: 0 2rem;
+
+  @media screen and (max-width: 45.999999em) {
+    margin-left: 0;
+    margin-right: 0;
+  }
+}
+
 .gurbani-word {
   position: relative;
